Revert 0.8 API changes for release

+++ tests/test_checkpointing.py
@@ -47,8 +47,7 @@ def test_read_write_P_2D(read_comm, family, degree, complex, mesh_2D):
                            degree,
                            basix.LagrangeVariant.gll_warped,
                            gdim=mesh.geometry.dim,
-                           shape=(mesh.geometry.dim, ),
-                           dtype=mesh.geometry.x.dtype)
+                           shape=(mesh.geometry.dim, ))
 
     def f(x):
         values = np.empty((2, x.shape[1]), dtype=f_dtype)
@@ -101,8 +100,7 @@ def test_read_write_P_2D_time(read_comm, family, degree, complex, mesh_2D):
                            degree,
                            basix.LagrangeVariant.gll_warped,
                            gdim=mesh.geometry.dim,
-                           shape=(mesh.geometry.dim, ),
-                           dtype=mesh.geometry.x.dtype)
+                           shape=(mesh.geometry.dim, ))
 
     def f0(x):
         values = np.empty((2, x.shape[1]), dtype=f_dtype)
